chore: rename all component files to `index.js`
- Rename all primary component files to `index.js` for consistency.
- Update `./client/src/components/index.js` accordingly
