+++ client/src/components/index.js
@@ -1,8 +1,8 @@
 export { default as CarouselItem } from './CarouselItem';
 export { default as FileUpload } from './FileUpload';
-export { default as Login } from './Login/Login';
+export { default as Login } from './Login';
 export { default as Main } from './Main';
-export { default as Nav } from './Nav/Nav';
+export { default as Nav } from './Nav';
 export { default as PlaceCard } from './PlaceCard';
 export { default as ResourceCard } from './ResourceCard';
-export { default as SignUp } from './SignUp/SignUp';
+export { default as SignUp } from './SignUp';
